W-16540082: stop serializing AVRORawSchema annotation
