+++ amf-cli/shared/src/test/resources/avro/tck/apis/invalid/avro-invalid.jsonld
@@ -155,21 +155,6 @@
                           }
                         ]
                       }
-                    ],
-                    "http://a.ml/vocabularies/document-source-maps#avro-raw-schema": [
-                      {
-                        "@id": "file://amf-cli/shared/src/test/resources/avro/tck/apis/invalid/avro-invalid.yaml#/declares/msg/testMessage/payload/default/any/default-any/source-map/avro-raw-schema/element_0",
-                        "http://a.ml/vocabularies/document-source-maps#element": [
-                          {
-                            "@value": "file://amf-cli/shared/src/test/resources/avro/tck/apis/invalid/avro-invalid.yaml#/declares/msg/testMessage/payload/default/any/default-any"
-                          }
-                        ],
-                        "http://a.ml/vocabularies/document-source-maps#value": [
-                          {
-                            "@value": "{\n  \"name\": \"Person\",\n  \"type\": \"nonexistent\"\n}"
-                          }
-                        ]
-                      }
                     ]
                   }
                 ]
