+++ amf-cli/shared/src/test/resources/upanddown/cycle/avro/valid/array.jsonld
@@ -53,21 +53,6 @@
                       }
                     ]
                   }
-                ],
-                "http://a.ml/vocabularies/document-source-maps#avro-raw-schema": [
-                  {
-                    "@id": "file://amf-cli/shared/src/test/resources/upanddown/cycle/avro/valid/array.json#/array/default-array/scalar/default-scalar/source-map/avro-raw-schema/element_0",
-                    "http://a.ml/vocabularies/document-source-maps#element": [
-                      {
-                        "@value": "file://amf-cli/shared/src/test/resources/upanddown/cycle/avro/valid/array.json#/array/default-array/scalar/default-scalar"
-                      }
-                    ],
-                    "http://a.ml/vocabularies/document-source-maps#value": [
-                      {
-                        "@value": "\"string\""
-                      }
-                    ]
-                  }
                 ]
               }
             ]
@@ -107,21 +92,6 @@
                 "http://a.ml/vocabularies/document-source-maps#value": [
                   {
                     "@value": "array"
-                  }
-                ]
-              }
-            ],
-            "http://a.ml/vocabularies/document-source-maps#avro-raw-schema": [
-              {
-                "@id": "file://amf-cli/shared/src/test/resources/upanddown/cycle/avro/valid/array.json#/array/default-array/source-map/avro-raw-schema/element_0",
-                "http://a.ml/vocabularies/document-source-maps#element": [
-                  {
-                    "@value": "file://amf-cli/shared/src/test/resources/upanddown/cycle/avro/valid/array.json#/array/default-array"
-                  }
-                ],
-                "http://a.ml/vocabularies/document-source-maps#value": [
-                  {
-                    "@value": "{\n  \"type\": \"array\",\n  \"items\": \"string\",\n  \"default\": []\n}"
                   }
                 ]
               }
