+++ amf-cli/shared/src/test/resources/upanddown/cycle/avro/valid/map.jsonld
@@ -53,21 +53,6 @@
                       }
                     ]
                   }
-                ],
-                "http://a.ml/vocabularies/document-source-maps#avro-raw-schema": [
-                  {
-                    "@id": "file://amf-cli/shared/src/test/resources/upanddown/cycle/avro/valid/map.json#/shape/default-node/scalar/default-scalar/source-map/avro-raw-schema/element_0",
-                    "http://a.ml/vocabularies/document-source-maps#element": [
-                      {
-                        "@value": "file://amf-cli/shared/src/test/resources/upanddown/cycle/avro/valid/map.json#/shape/default-node/scalar/default-scalar"
-                      }
-                    ],
-                    "http://a.ml/vocabularies/document-source-maps#value": [
-                      {
-                        "@value": "\"long\""
-                      }
-                    ]
-                  }
                 ]
               }
             ]
@@ -105,21 +90,6 @@
                 "http://a.ml/vocabularies/document-source-maps#value": [
                   {
                     "@value": "map"
-                  }
-                ]
-              }
-            ],
-            "http://a.ml/vocabularies/document-source-maps#avro-raw-schema": [
-              {
-                "@id": "file://amf-cli/shared/src/test/resources/upanddown/cycle/avro/valid/map.json#/shape/default-node/source-map/avro-raw-schema/element_0",
-                "http://a.ml/vocabularies/document-source-maps#element": [
-                  {
-                    "@value": "file://amf-cli/shared/src/test/resources/upanddown/cycle/avro/valid/map.json#/shape/default-node"
-                  }
-                ],
-                "http://a.ml/vocabularies/document-source-maps#value": [
-                  {
-                    "@value": "{\n  \"type\": \"map\",\n  \"values\": \"long\",\n  \"default\": {}\n}"
                   }
                 ]
               }
