+++ amf-cli/shared/src/test/resources/upanddown/cycle/avro/valid/schema-in-field.jsonld
@@ -82,21 +82,6 @@
                                   }
                                 ]
                               }
-                            ],
-                            "http://a.ml/vocabularies/document-source-maps#avro-raw-schema": [
-                              {
-                                "@id": "file://amf-cli/shared/src/test/resources/upanddown/cycle/avro/valid/schema-in-field.json#/shape/AllTypes/property/property/default-property/array/default-array/array/default-array/scalar/default-scalar/source-map/avro-raw-schema/element_0",
-                                "http://a.ml/vocabularies/document-source-maps#element": [
-                                  {
-                                    "@value": "file://amf-cli/shared/src/test/resources/upanddown/cycle/avro/valid/schema-in-field.json#/shape/AllTypes/property/property/default-property/array/default-array/array/default-array/scalar/default-scalar"
-                                  }
-                                ],
-                                "http://a.ml/vocabularies/document-source-maps#value": [
-                                  {
-                                    "@value": "\"string\""
-                                  }
-                                ]
-                              }
                             ]
                           }
                         ]
@@ -119,21 +104,6 @@
                             "http://a.ml/vocabularies/document-source-maps#value": [
                               {
                                 "@value": "array"
-                              }
-                            ]
-                          }
-                        ],
-                        "http://a.ml/vocabularies/document-source-maps#avro-raw-schema": [
-                          {
-                            "@id": "file://amf-cli/shared/src/test/resources/upanddown/cycle/avro/valid/schema-in-field.json#/shape/AllTypes/property/property/default-property/array/default-array/array/default-array/source-map/avro-raw-schema/element_0",
-                            "http://a.ml/vocabularies/document-source-maps#element": [
-                              {
-                                "@value": "file://amf-cli/shared/src/test/resources/upanddown/cycle/avro/valid/schema-in-field.json#/shape/AllTypes/property/property/default-property/array/default-array/array/default-array"
-                              }
-                            ],
-                            "http://a.ml/vocabularies/document-source-maps#value": [
-                              {
-                                "@value": "{\n  \"type\": \"array\",\n  \"items\": \"string\"\n}"
                               }
                             ]
                           }
@@ -179,21 +149,6 @@
                           }
                         ]
                       }
-                    ],
-                    "http://a.ml/vocabularies/document-source-maps#avro-raw-schema": [
-                      {
-                        "@id": "file://amf-cli/shared/src/test/resources/upanddown/cycle/avro/valid/schema-in-field.json#/shape/AllTypes/property/property/default-property/array/default-array/source-map/avro-raw-schema/element_0",
-                        "http://a.ml/vocabularies/document-source-maps#element": [
-                          {
-                            "@value": "file://amf-cli/shared/src/test/resources/upanddown/cycle/avro/valid/schema-in-field.json#/shape/AllTypes/property/property/default-property/array/default-array"
-                          }
-                        ],
-                        "http://a.ml/vocabularies/document-source-maps#value": [
-                          {
-                            "@value": "{\n  \"type\": \"array\",\n  \"items\": {\n    \"type\": \"array\",\n    \"items\": \"string\"\n  },\n  \"default\": []\n}"
-                          }
-                        ]
-                      }
                     ]
                   }
                 ]
@@ -221,21 +176,6 @@
                 "@id": "file://amf-cli/shared/src/test/resources/upanddown/cycle/avro/valid/schema-in-field.json#/shape/AllTypes/property/property/default-property/source-map",
                 "@type": [
                   "http://a.ml/vocabularies/document-source-maps#SourceMap"
-                ],
-                "http://a.ml/vocabularies/document-source-maps#avro-raw-schema": [
-                  {
-                    "@id": "file://amf-cli/shared/src/test/resources/upanddown/cycle/avro/valid/schema-in-field.json#/shape/AllTypes/property/property/default-property/source-map/avro-raw-schema/element_0",
-                    "http://a.ml/vocabularies/document-source-maps#element": [
-                      {
-                        "@value": "http://a.ml/vocabularies/shapes#range"
-                      }
-                    ],
-                    "http://a.ml/vocabularies/document-source-maps#value": [
-                      {
-                        "@value": "{\n  \"type\": \"array\",\n  \"items\": {\n    \"type\": \"array\",\n    \"items\": \"string\"\n  },\n  \"default\": []\n}"
-                      }
-                    ]
-                  }
                 ],
                 "http://a.ml/vocabularies/document-source-maps#avro-schema": [
                   {
@@ -306,21 +246,6 @@
                 "http://a.ml/vocabularies/document-source-maps#value": [
                   {
                     "@value": "record"
-                  }
-                ]
-              }
-            ],
-            "http://a.ml/vocabularies/document-source-maps#avro-raw-schema": [
-              {
-                "@id": "file://amf-cli/shared/src/test/resources/upanddown/cycle/avro/valid/schema-in-field.json#/shape/AllTypes/source-map/avro-raw-schema/element_0",
-                "http://a.ml/vocabularies/document-source-maps#element": [
-                  {
-                    "@value": "file://amf-cli/shared/src/test/resources/upanddown/cycle/avro/valid/schema-in-field.json#/shape/AllTypes"
-                  }
-                ],
-                "http://a.ml/vocabularies/document-source-maps#value": [
-                  {
-                    "@value": "{\n  \"type\": \"record\",\n  \"name\": \"AllTypes\",\n  \"namespace\": \"root\",\n  \"aliases\": [\n    \"EveryTypeInTheSameSchema\"\n  ],\n  \"doc\": \"this schema contains every possible type you can declare in avro inside it's fields\",\n  \"fields\": [\n    {\n      \"type\": \"array\",\n      \"items\": {\n        \"type\": \"array\",\n        \"items\": \"string\"\n      },\n      \"default\": []\n    }\n  ]\n}"
                   }
                 ]
               }
