+++ amf-shapes/shared/src/main/scala/amf/shapes/internal/validation/avro/BaseAvroSchemaPayloadValidator.scala
@@ -1,18 +1,13 @@
 package amf.shapes.internal.validation.avro
 
 import amf.core.client.common.validation._
-import amf.core.client.scala.errorhandling.UnhandledErrorHandler
-import amf.core.client.scala.model.DataType
 import amf.core.client.scala.model.document.PayloadFragment
 import amf.core.client.scala.model.domain._
-import amf.core.client.scala.parse.document.SyamlParsedDocument
 import amf.core.client.scala.validation.payload.{PayloadParsingResult, ShapeValidationConfiguration}
 import amf.core.client.scala.validation.{AMFValidationReport, AMFValidationResult}
-import amf.core.internal.plugins.syntax.SyamlSyntaxRenderPlugin
 import amf.core.internal.remote.Mimes._
 import amf.shapes.client.scala.model.domain.AnyShape
 import amf.shapes.internal.annotations.AVRORawSchema
-import amf.shapes.internal.spec.common.emitter.PayloadEmitter
 import amf.shapes.internal.validation.common.{
   CommonBaseSchemaPayloadValidator,
   ReportValidationProcessor,
@@ -20,7 +15,6 @@ import amf.shapes.internal.validation.common.{
 }
 import amf.shapes.internal.validation.definitions.ShapePayloadValidations.ExampleValidationErrorSpecification
 
-import java.io.StringWriter
 import scala.concurrent.{ExecutionContext, Future}
 
 class ExampleUnknownException(e: Throwable) extends RuntimeException(e)
@@ -65,7 +59,6 @@ abstract class BaseAvroSchemaPayloadValidator(
       fragment: PayloadFragment,
       validationProcessor: ValidationProcessor
   ): AMFValidationReport = {
-
     try {
       performValidation(buildCandidate(fragment), validationProcessor)
     } catch {
@@ -103,29 +96,8 @@ abstract class BaseAvroSchemaPayloadValidator(
     }
   }
 
-  protected def literalRepresentation(payload: PayloadFragment): Option[String] = {
-    val futureText = payload.raw match {
-      case Some("") => None
-      case _ =>
-        val document = PayloadEmitter(payload.encodes)(UnhandledErrorHandler).emitDocument()
-        val writer   = new StringWriter()
-        SyamlSyntaxRenderPlugin.emit(`application/json`, SyamlParsedDocument(document), writer).map(_.toString)
-    }
-
-    futureText map { text =>
-      payload.encodes match {
-        case node: ScalarNode
-            if node.dataType
-              .option()
-              .contains(DataType.String) && text.nonEmpty && text.head != '"' =>
-          "\"" + text.stripLineEnd + "\""
-        case _ => text.stripLineEnd
-      }
-    }
-  }
-
   private def getRawAvroSchema(shape: Shape, validationProcessor: ValidationProcessor): Option[CharSequence] = {
-    val raw = shape.annotations.find(classOf[AVRORawSchema]).map(_.value)
+    val raw = shape.annotations.find(classOf[AVRORawSchema]).map(_.avroRawSchema)
     if (raw.isEmpty) {
       // TODO complete the result to make it a valid violation that the shape doesn't have the raw annotation
       // emitirlo?
