+++ amf-shapes/shared/src/main/scala/amf/shapes/internal/validation/common/CommonBaseSchemaPayloadValidator.scala
@@ -2,12 +2,19 @@ package amf.shapes.internal.validation.common
 
 import amf.core.client.common.validation.ProfileNames.AMF
 import amf.core.client.common.validation.{ProfileName, SeverityLevels, ValidationMode}
+import amf.core.client.scala.errorhandling.UnhandledErrorHandler
+import amf.core.client.scala.model.DataType
 import amf.core.client.scala.model.document.PayloadFragment
-import amf.core.client.scala.model.domain.DomainElement
+import amf.core.client.scala.model.domain.{DomainElement, ScalarNode}
+import amf.core.client.scala.parse.document.SyamlParsedDocument
 import amf.core.client.scala.validation.{AMFValidationReport, AMFValidationResult}
 import amf.core.client.scala.validation.payload.AMFShapePayloadValidator
+import amf.core.internal.plugins.syntax.SyamlSyntaxRenderPlugin
+import amf.core.internal.remote.Mimes.`application/json`
+import amf.shapes.internal.spec.common.emitter.PayloadEmitter
 import amf.shapes.internal.validation.definitions.ShapePayloadValidations.ExampleValidationErrorSpecification
 
+import java.io.StringWriter
 import scala.collection.mutable
 
 abstract class CommonBaseSchemaPayloadValidator extends AMFShapePayloadValidator {
@@ -51,4 +58,24 @@ abstract class CommonBaseSchemaPayloadValidator extends AMFShapePayloadValidator
       validationProcessor: ValidationProcessor
   ): Either[AMFValidationReport, Option[LoadedSchema]]
 
+  protected def literalRepresentation(payload: PayloadFragment): Option[String] = {
+    val futureText = payload.raw match {
+      case Some("") => None
+      case _ =>
+        val document = PayloadEmitter(payload.encodes)(UnhandledErrorHandler).emitDocument()
+        val writer   = new StringWriter()
+        SyamlSyntaxRenderPlugin.emit(`application/json`, SyamlParsedDocument(document), writer).map(_.toString)
+    }
+
+    futureText map { text =>
+      payload.encodes match {
+        case node: ScalarNode
+            if node.dataType
+              .option()
+              .contains(DataType.String) && text.nonEmpty && text.head != '"' =>
+          "\"" + text.stripLineEnd + "\""
+        case _ => text.stripLineEnd
+      }
+    }
+  }
 }
