+++ amf-shapes/shared/src/main/scala/amf/shapes/internal/validation/jsonschema/BaseJsonSchemaPayloadValidator.scala
@@ -3,7 +3,7 @@ package amf.shapes.internal.validation.jsonschema
 import amf.core.client.common.render.JsonSchemaDraft7
 import amf.core.client.common.validation._
 import amf.core.client.scala.config.{ParsingOptions, RenderOptions}
-import amf.core.client.scala.errorhandling.{AMFErrorHandler, UnhandledErrorHandler}
+import amf.core.client.scala.errorhandling.AMFErrorHandler
 import amf.core.client.scala.model.DataType
 import amf.core.client.scala.model.document.PayloadFragment
 import amf.core.client.scala.model.domain._
@@ -17,7 +17,6 @@ import amf.core.internal.plugins.syntax.{SYamlAMFParserErrorHandler, SyamlSyntax
 import amf.core.internal.remote.Mimes._
 import amf.core.internal.validation.core.ValidationSpecification
 import amf.shapes.client.scala.model.domain.{AnyShape, FileShape, ScalarShape, UnionShape}
-import amf.shapes.internal.spec.common.emitter.PayloadEmitter
 import amf.shapes.internal.spec.jsonschema.emitter.JsonSchemaEmitter
 import amf.shapes.internal.validation.common.{
   CommonBaseSchemaPayloadValidator,
@@ -138,27 +137,6 @@ abstract class BaseJsonSchemaPayloadValidator(
     validationProcessor.keepResults(eh.getResults)
     val writer = new StringWriter()
     SyamlSyntaxRenderPlugin.emit(`application/json`, document, writer).map(_.toString)
-  }
-
-  protected def literalRepresentation(payload: PayloadFragment): Option[String] = {
-    val futureText = payload.raw match {
-      case Some("") => None
-      case _ =>
-        val document = PayloadEmitter(payload.encodes)(UnhandledErrorHandler).emitDocument()
-        val writer   = new StringWriter()
-        SyamlSyntaxRenderPlugin.emit(`application/json`, SyamlParsedDocument(document), writer).map(_.toString)
-    }
-
-    futureText map { text =>
-      payload.encodes match {
-        case node: ScalarNode
-            if node.dataType
-              .option()
-              .contains(DataType.String) && text.nonEmpty && text.head != '"' =>
-          "\"" + text.stripLineEnd + "\""
-        case _ => text.stripLineEnd
-      }
-    }
   }
 
   private def getOrCreateSchema(
